Viewer_1-master/requirements.txt
alembic==1.0.8
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Dance==1.2.0
Flask-Extension==1.0
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-MySQL==1.4.0
Flask-Session==0.3.1
Flask-SocketIO==3.1.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
lazy==1.4
Mako==1.0.7
MarkupSafe==1.1.1
oauthlib==2.1.0
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.4.3
python-socketio==3.1.2
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.0
urllib3==1.24.1
URLObject==2.4.3
Werkzeug==0.14.1
WTForms==2.2.1
